DO NOT MERGE Speculative fix for BT race condition
When the BT stack goes down very soon after a device added broadcast is
received, we have the potential for a device to stick around in the
device manager with no active BT service, leading to repeated failure to
connect to said device. This change checks to make sure the service is
non null before adding the device.

Change-Id: Ibee8ba2c611d750147249ce1c750a8d433ebff5b
Bug: 124235726
Test: unit